+++ lib/perilune.rb
@@ -2,7 +2,17 @@
 
 require 'perilune/engine'
 require 'perilune/tasks/mixin'
+require 'perilune/configuration'
 
 module Perilune
   # Your code goes here...
+  def self.default
+    @default ||= Configuration.new
+  end
+
+  def self.configure
+    yield(default)
+
+    default
+  end
 end