+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Feature: Custom configuration for Trifle:Stat and Cron Job Queue Name
+
 - Feature: Migration and Generater added for create the required table
 
 - Feature: Added Trifle::Stat for analytics